+++ core/components/romanesco/elements/plugins/07_computations/c_content/manipulatedom.plugin.php
@@ -66,6 +66,7 @@ switch ($modx->event->name) {
                     '.divider:not(.hidden)',
                     '.basic.form',
                     '.accordion:not(.styled)',
+                    '.text.menu',
                 );
 
                 // Revert inverted styling inside these nested elements