+++ core/components/romanesco/docs/changelog.md
@@ -5,8 +5,11 @@ Released on ...
 
 New features:
 - Add post hook for pThumb to optimize the image and generate a WebP version
+- Add ContentBlock with Table of Contents menu
 
 Fixes and improvements:
+- Use placeholder to detect ToC instead of toc_enabled_templates system setting
+- Show ToC submenu in dropdown on tablet and mobile
 - Move shared overview settings to separate molecules
 - Fix regClient scripts not being parsed in Markdown pages on first load
 - Load external JS from head with defer